fix: Fix some incorrect usages of `tComponent`
